+++ src/stores/SentryStore/SentryStore.ts
@@ -1,130 +1,54 @@
-import * as Sentry from '@sentry/react';
-
-import { logError } from '../../utils';
+import { User } from '@sentry/types';
+import { BrowserOptions } from '@sentry/browser/dist/backend';
 
 import {
-  SentryConfigType,
-  SentryTagEnum,
-  SentryUserType,
-  ApiErrorDataType,
-  ExceptionType,
+  captureAPIException,
+  captureException,
+  captureFAPIException,
+  captureRenderException,
+  captureVKBridgeException,
+  init,
+  setUser,
+} from './utils';
+import {
+  APIErrorDataType,
   RenderExceptionType,
-  ApiExceptionType,
+  APIExceptionType,
   VKBridgeExceptionType,
+  FAPIExceptionType,
+  ExceptionType,
 } from './types';
 
 class SentryStore {
-  private _inited = false;
-
-  get shouldSendError(): boolean {
-    return Boolean(
-      window.is_production && Sentry.getCurrentHub().getClient() && this._inited
-    );
+  init(options: BrowserOptions, user?: User): void {
+    init(options, user);
   }
 
-  private _setInited = (value: boolean): void => {
-    this._inited = value;
-  };
+  setUser(user: User): void {
+    setUser(user);
+  }
 
-  private _captureException = (
-    ...args: Parameters<typeof Sentry.captureException>
-  ): void => {
-    if (this.shouldSendError) {
-      Sentry.captureException(...args);
-    }
-  };
+  captureException<E = any>(exception: ExceptionType<E>): void {
+    captureException<E>(exception);
+  }
 
-  init = ({
-    dsn = process.env.SENTRY_DSN,
-    normalizeDepth = 6,
-    user,
-  }: SentryConfigType): void => {
-    if (!window.is_production) {
-      return;
-    }
+  captureRenderException<E = any>(exception: RenderExceptionType<E>): void {
+    captureRenderException<E>(exception);
+  }
 
-    if (!dsn) {
-      logError('DSN argument was not provided for Sentry');
-      return;
-    }
+  captureAPIException<ED extends APIErrorDataType = APIErrorDataType, E = any>(
+    exception: APIExceptionType<ED, E>
+  ): void {
+    captureAPIException<ED, E>(exception);
+  }
 
-    Sentry.init({
-      dsn,
-      normalizeDepth,
-    });
+  captureVKBridgeException<E = any>(exception: VKBridgeExceptionType<E>): void {
+    captureVKBridgeException<E>(exception);
+  }
 
-    if (user) {
-      this.setUser(user);
-    }
-
-    this._setInited(true);
-  };
-
-  setUser = (config: SentryUserType): void => {
-    Sentry.setUser(config);
-  };
-
-  captureException = <T = any>({
-    error,
-    extra,
-    type = SentryTagEnum.custom,
-  }: ExceptionType<T>): void => {
-    this._captureException(error, {
-      tags: {
-        type,
-      },
-      extra,
-    });
-  };
-
-  captureRenderException = <T = any>({
-    error,
-    extra,
-  }: RenderExceptionType<T>): void => {
-    this._captureException(error, {
-      tags: {
-        type: SentryTagEnum.render,
-      },
-      extra,
-    });
-  };
-
-  captureApiException = <ED extends ApiErrorDataType, T = any>({
-    error,
-    errorData,
-    endpoint,
-    payload,
-  }: ApiExceptionType<T, ED>): void => {
-    this._captureException(error, {
-      tags: {
-        type: SentryTagEnum.api,
-        url: endpoint.url,
-        code: errorData.code,
-      },
-      extra: {
-        payload,
-        errorData,
-      },
-    });
-  };
-
-  captureVKBridgeException = ({
-    errorEvent,
-    url,
-    payload,
-  }: VKBridgeExceptionType): void => {
-    this._captureException(errorEvent.type, {
-      tags: {
-        type: SentryTagEnum.vkBridge,
-        vkUrl: url,
-        vkCode: errorEvent.type,
-      },
-      extra: {
-        payload,
-        errorData: errorEvent.data,
-      },
-    });
-  };
+  captureFAPIException<E = any>(exception: FAPIExceptionType<E>): void {
+    captureFAPIException<E>(exception);
+  }
 }
 
 export default SentryStore;