+++ src/stores/SentryStore/index.ts
@@ -1,3 +1,5 @@
 export { default as SentryStore } from './SentryStore';
 
 export * from './types';
+
+export * from './utils';